meta: Exclude implementation-internal signals in sigmasks() by default
Now, unless include_internal=True is passed, any signals that:

a) Don't have entries in signal.Signals, and
b) Aren't in the range [SIGRTMIN, SIGRTMAX] (this is assumed true if
   SIGRTMIN/SIGRTMAX aren't defined)

are excluded from the returned signal sets. This means that any extra
signals that might be used specially by the libc are not exposed to end
users. (That information is, however, still available if it is needed.)