_util: Add new read_file_first_line() helper
On Linux, some special files in /proc and /sys will only ever have one
line. This function saves a syscall if the first read() call is able to
get the entire contents (read_file() has to make a second read() call
because for all it knows the first one might have been a short read).