pypsutil: Add BatteryInfo.status; rework BatteryInfo.power_plugged
Now it should be more correct for systems with multiple batteries.

In the BatteryInfo objects returned by sensors_power(), power_plugged is
True if the battery is actively charging and False if the battery is
actively discharging; in all other cases it is None.

In the BatteryInfo object returned by sensors_battery(), power_plugged
may be set based on the reported AC adapter status. This is because
sensors_battery() is intended to be a general "system-wide power
information" function. (Though it's obviously incorrect if there are
multiple batteries.)

sensors_is_on_ac_power() is also now more correct.