meta: Make create_time() more reliable on Linux
create_time() relies on boot_time(), which is reliant on time.time(),
which is affected by NTP updates. As a result, an NTP update could skew
the values returned by create_time(), which breaks fundamental
assumptions of this library.

Now each Process has a create_time and a raw_create_time. create_time is
the creation time in seconds since the Unix epoch, while raw_create_time
is the value returned by the OS (and it's what is used for process
identity comparisons).

On all platforms other than Linux, these are the same thing, but on
Linux it's an important distinction.